Refactor to create AbstractPlatformClient
Common abstract class for PlatformClient implementations to extend for shared functionality such as notifications on async target discovery
