+++ src/main/java/com/redhat/rhjmc/containerjfr/platform/openshift/OpenShiftPlatformStrategy.java
@@ -39,6 +39,8 @@
  * SOFTWARE.
  * #L%
  */
+
+//FIXME move the openshift package under the internal package
 package com.redhat.rhjmc.containerjfr.platform.openshift;
 
 import java.nio.file.Files;
